optimsied memory for memory layers

+++ src/memory_layer.py
@@ -31,13 +31,13 @@ class MemoryPlusLayer(nn.Module):
         
         # Sub-Key Matrix One and Two
         # NOTE: Don't use nn.linear here, due to sparse key retrieval mechanism in forward pass.
-        self.subkey_one = nn.Parameter(torch.empty(self.num_subkeys, self.subkey_dim))
-        self.subkey_two = nn.Parameter(torch.empty(self.num_subkeys, self.subkey_dim))
+        self.subkey_one = nn.Parameter(torch.empty(self.num_subkeys, self.subkey_dim, dtype=torch.float32))
+        self.subkey_two = nn.Parameter(torch.empty(self.num_subkeys, self.subkey_dim, dtype=torch.float32))
         nn.init.uniform_(self.subkey_one, a = -1, b = 1)
         nn.init.uniform_(self.subkey_two, a = -1, b = 1)
 
         # Value Matrix
-        self.values = nn.Parameter(torch.empty(memory_slots, self.value_dim))
+        self.values = nn.Parameter(torch.empty(memory_slots, self.value_dim, dtype=torch.float32))
         nn.init.normal_(self.values, std=0.02)  # apparently from lample et al 2019, CAN't FIND ITS REFERENCE
 
         # Weight Matrix One
